refactor(cli): Splitted cli-parser in smaller modules
