+++ src/cli-parser/get-plugins.js
@@ -1,0 +1,25 @@
+const path = require('path')
+const fs = require('fs')
+
+module.exports = (confFile, pluginsList) => {
+  let res
+  if (confFile) {
+    // get full path to config file
+    const composerPath = path.join(process.cwd(), confFile)
+    try {
+      // check file exists
+      fs.accessSync(composerPath, fs.F_OK)
+      // access the module and retrieve the list of plugins
+      res = require(composerPath)()
+    } catch (e) {
+      res = `Module '${composerPath}' doesn't exist or isn't a valid module.`
+    }
+  } else if (pluginsList) {
+    try {
+      res = pluginsList.map((moduleName) => require(moduleName)())
+    } catch (e) {
+      res = e ? e.message : `Something wrong happened requiring postcss plugins`
+    }
+  }
+  return res
+}
